requirements.txt
easyocr
streamlit
opencv-python-headless
numpy
pandas
openai
python-dotenv
